+++ NetherGate/src/com/elmakers/mine/bukkit/plugins/nether/BlockRequest.java
@@ -10,7 +10,7 @@ import org.bukkit.util.BlockVector;
 import org.bukkit.util.Vector;
 
 import com.elmakers.mine.bukkit.plugins.nether.dao.NetherWorld;
-import com.elmakers.mine.bukkit.plugins.nether.listener.BlockRequestListener;
+import com.elmakers.mine.bukkit.plugins.persistence.BlockRequestListener;
 import com.elmakers.mine.bukkit.plugins.persistence.dao.BoundingBox;
 
 public class BlockRequest